refactor: consolidate tool management and improve imports
Major refactoring of tool management and utility functions:
- Consolidated tool execution logic into ToolManager class
- Created new toolFormatters.ts for pure formatting functions
- Removed duplicate client initialization in demo
- Cleaned up import statements (removed .js extensions)
- Removed deprecated toolHelpers.ts
- Added proper getter method for clients in ToolManager
- Improved encapsulation of tool-related functionality

This refactoring improves code organization, reduces duplication, and makes the codebase more maintainable.
