--- test-files/test.txt
@@ -1,1 +1,0 @@
-poopoo peepepe
